Move pyoidide_testing_app to dummyserver.app
It should be able to share more code with hypercorn_app
